Rename PropertyUpdaterService to DatabaseService & added query() method.
Rename DatabaseService to DeprecatedDBService.
