+++ src/core/index.ts
@@ -1,6 +1,6 @@
 export * from './logger';
 export * from './config/config.service';
 export * from './core.module';
-export * from './database.service';
-export * from './database/property-updater.service';
+export * from './deprecated-database.service';
+export * from './database/database.service';
 export * from './database/indexer';
